move lb to file
